Move both event listeners to the same class
